validation-gen: split the generator into importable packages
validation-gen was a single main package, so nothing in it could be
imported. A project that wants this generator with its own tags, or with
a different tag prefix, has to be able to build its own binary on the
same code.

Move the generator into k8s.io/code-generator/cmd/validation-gen/generators
and its flags into .../args, matching the layout of defaulter-gen and
conversion-gen. printDocs becomes generators.PrintDocs, which writes to an
io.Writer instead of stdout. main.go is now just flag parsing and a call to
gengo.Execute.

No behavior change: the generated output and --docs output are byte
identical, modulo the trailing newline that fmt.Println added to --docs.
