+++ staging/src/k8s.io/code-generator/cmd/validation-gen/generators/lint_rules.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package generators
 
 import (
 	"fmt"
